requirements.txt
pyside6==6.2.0
sqlalchemy==1.4.25
argon2-cffi==21.1.0
